+++ src/web.ts
@@ -527,7 +527,6 @@ export class SocialLoginWeb extends WebPlugin implements SocialLoginPlugin {
               token: "", // TODO: to fix and find the correct token
             },
             idToken: res.authorization.id_token || null,
-            authorizationCode: res.authorization.code || null,
           };
           resolve({ provider: "apple", result });
         })
@@ -603,7 +602,7 @@ export class SocialLoginWeb extends WebPlugin implements SocialLoginPlugin {
                     hometown: null,
                     profileURL: null,
                   },
-                  authenticationToken: null,
+                  idToken: null,
                 };
                 resolve({ provider: "facebook", result });
               },